Hopper Dropdown Menus + Code Improvements
Added Managed Tasks. Managed Tasks hold a reference to an active task, can cancel it, or add a new task, but will only allow a single task to exist at any one time. Managed tasks can also cancel a task that a rover is currently executing, interrupting the rover and forcing it to search for a new task.